Revert "[TI-5] Fix sing up"
