+++ src/Components/Header/styles.module.scss
@@ -10,6 +10,7 @@
   top: 0;
   z-index: 1;
 
+  
   nav {
     width: 100%;
     max-width: 1200px;
@@ -17,6 +18,9 @@
     align-items: center;
     justify-content: space-between;
 
+    img{
+      height: 170px;
+    }
   }
 
   .login {
@@ -47,21 +51,13 @@
 }
 
 .options{
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: row;
-  justify-content: flex-start;
+  display: flex; 
+  justify-content: space-evenly;
   width: 50%;
 
-  div{
-    float: left;
-    width: 33.3%;
-    box-sizing: border-box;
-
-    span{
-      color: #F7F8FA;
-      font: 500 1rem Roboto, sans-serif;
-      cursor: pointer;
-    }
+  span{
+    color: #F7F8FA;
+    font: 500 1rem Roboto, sans-serif;
+    cursor: pointer;
   }
 }
