--- src/Components/Layout/Footer/styles.module.scss
@@ -1,66 +1,0 @@
-.footer {
-	display: flex;
-	justify-content: center;
-	width: 100%;
-	height: 320px;
-	-webkit-box-pack: center;
-	background-color: #006D77;
-	bottom: 0;
-	z-index: 1000;
-  
-	footer {
-	  max-width: 1200px;
-	  width: 100%;
-	  padding: 0 2rem;
-	  color: #FFF;
-	  display: flex;
-	  flex-wrap: wrap;
-	  justify-content: space-between;
-	  align-content: flex-start;
-	}
-  }
-  
-  .line {
-	display: block;
-	margin-top: 3rem;
-	overflow: hidden;
-	border-style: inset;
-	border-width: 0.3px;
-	width: 100%;
-	height: 1px;
-	background-color: #ffffff;
-	margin-bottom: 2rem;
-  }
-  
-  .logo {
-	display: flex;
-	flex-direction: column;
-  
-	img {
-	  width: 200px;
-	}
-  
-	span {
-	  font: 500 1rem Poppins, sans-serif;
-	}
-  }
-  
-  .options {
-	display: flex;
-  
-	span {
-	  text-decoration: underline;
-	  font: 500 1rem Roboto, sans-serif;
-  
-	}
-  
-	ul {
-	  list-style-type: none;
-	}
-  
-	li {
-	  text-decoration: none;
-	  font: 300 1rem Roboto, sans-serif;
-	  padding: 0.5rem 0;
-	}
-  }
