--- src/Components/Layout/default.tsx
@@ -1,14 +1,0 @@
-import Header from './Header'
-import Footer from './Footer'
-
-export default function DefaultLayout({ children }) {
-
-  return (
-    <>
-      <Header />
-      <main>{children}</main>
-      <Footer />
-    </>
-  )
-  
-}
