+++ src/Components/Orders/index.tsx
@@ -5,7 +5,7 @@ import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
-import Title from 'Common/Title';
+import Title from '../Title';
 
 // Generate Order Data
 function createData(
