+++ src/Components/Orders/styles.module.scss
@@ -1,0 +1,3 @@
+.seeMore {
+    marginTop: theme.spacing(3),
+}
