--- src/Components/Site/Banner/index.tsx
@@ -1,8 +1,0 @@
-import styles from './styles.module.scss';
-
-export function Banner() {
-  return (
-    <div className={styles.banner}>
-    </div>
-  )
-}
