--- src/Components/Site/Banner/styles.module.scss
@@ -1,10 +1,0 @@
-.banner {
-  width: 100%;
-  height: 600px;
-  position: absolute;
-  z-index: -1;
-
-  background-image: url("/assets/bg-image.png");
-  background-repeat: no-repeat;
-  background-size: cover;
-}
