--- src/Components/common/Container/index.tsx
@@ -1,13 +1,0 @@
-import styles from './styles.module.scss';
-
-type ContainerProps = {
-  children: any
-}
-
-export function Container(props: ContainerProps) {
-  return (
-    <div className={styles.container}>
-      {props.children}
-    </div>
-  )
-}
