--- src/Components/common/Container/styles.module.scss
@@ -1,8 +1,0 @@
-.container {
-  width: 100%;
-  max-width: 1200px;
-  min-height: calc(100vh - 440px);
-  justify-content: center;
-  -webkit-box-pack: center;
-  margin: 0 auto;
-}
