+++ src/pages/Dashboard/places.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
-import { ReactElement } from 'react';
-import {Container, Typography, Box, Button} from '@material-ui/core';
-import Link from 'Common/Link';
-import DashLayout from 'Components/Layout/dash';
+import Container from '@material-ui/core/Container';
+import Typography from '@material-ui/core/Typography';
+import Box from '@material-ui/core/Box';
+import Button from '@material-ui/core/Button';
+import Link from '../../Components/Link';
 
 export default function Places() {
   return (
@@ -20,10 +21,4 @@ export default function Places() {
       </Box>
     </Container>
   );
-}
-
-Places.getLayout = function getLayout(page: ReactElement) {
-  return (
-    <DashLayout>{page}</DashLayout>
-  )
 }
