+++ src/pages/Site/about.tsx
@@ -1,6 +1,9 @@
 import * as React from 'react';
-import {Container, Typography, Box, Button} from '@material-ui/core';
-import Link from 'Common/Link';
+import Container from '@material-ui/core/Container';
+import Typography from '@material-ui/core/Typography';
+import Box from '@material-ui/core/Box';
+import Button from '@material-ui/core/Button';
+import Link from '../../Components/Link';
 
 export default function About() {
   return (
