+++ src/pages/Site/places.tsx
@@ -1,6 +1,15 @@
 import * as React from 'react';
-import {CssBaseline, Typography, Box, Container, Grid, Stack, Button} from '@material-ui/core';
-import {Card, CardActions, CardContent, CardMedia} from '@material-ui/core';
+import Button from '@material-ui/core/Button';
+import Card from '@material-ui/core/Card';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import Grid from '@material-ui/core/Grid';
+import Stack from '@material-ui/core/Stack';
+import Box from '@material-ui/core/Box';
+import Typography from '@material-ui/core/Typography';
+import Container from '@material-ui/core/Container';
 
 const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9];
