+++ src/pages/Site/signUp.tsx
@@ -1,10 +1,14 @@
 import * as React from 'react';
+import Avatar from '@material-ui/core/Avatar';
 import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import TextField from '@material-ui/core/TextField';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Checkbox from '@material-ui/core/Checkbox';
 import Link from '@material-ui/core/Link';
 import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
+import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
 import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
 
@@ -24,14 +28,17 @@ export default function SignUp() {
       <CssBaseline />
       <Box
         sx={{
-          marginTop: 10,
+          marginTop: 8,
           display: 'flex',
           flexDirection: 'column',
           alignItems: 'center',
         }}
       >
+        <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+          <LockOutlinedIcon />
+        </Avatar>
         <Typography component="h1" variant="h5">
-          Área de Cadastro
+          Sign up
         </Typography>
         <Box component="form" noValidate onSubmit={handleSubmit} sx={{ mt: 3 }}>
           <Grid container spacing={2}>
@@ -42,7 +49,7 @@ export default function SignUp() {
                 required
                 fullWidth
                 id="firstName"
-                label="Nome"
+                label="First Name"
                 autoFocus
               />
             </Grid>
@@ -51,161 +58,46 @@ export default function SignUp() {
                 required
                 fullWidth
                 id="lastName"
-                label="Sobrenome"
+                label="Last Name"
                 name="lastName"
                 autoComplete="lname"
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                id="email"
-                label="Email"
-                name="email"
-                autoComplete="email"
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                id="cpf"
-                label="CPF"
-                name="cpf"
-                autoComplete="cpf"
-              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <TextField
-                required
-                fullWidth
-                id="gender"
-                label="Gênero"
-                name="gender"
-                autoComplete="gender"
-              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <TextField
-                required
-                fullWidth
-                id="phone"
-                label="Telefone"
-                name="phone"
-                autoComplete="phone"
-              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <TextField
-                required
-                fullWidth
-                id="dateBirth"
-                label="Data de Nascimento"
-                name="dateBirth"
-                autoComplete="dateBirth"
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                id="cep"
-                label="CEP"
-                name="cep"
-                autoComplete="cep"
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                id="estado"
-                label="Estado"
-                name="estado"
-                autoComplete="estado"
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                id="city"
-                label="Cidade"
-                name="city"
-                autoComplete="city"
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                id="district"
-                label="Bairro"
-                name="district"
-                autoComplete="district"
               />
             </Grid>
             <Grid item xs={12}>
               <TextField
                 required
                 fullWidth
-                id="street"
-                label="Rua"
-                name="street"
-                autoComplete="street"
+                id="email"
+                label="Email Address"
+                name="email"
+                autoComplete="email"
               />
             </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                id="number"
-                label="Número"
-                name="number"
-                autoComplete="number"
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                id="complement"
-                label="Complemento"
-                name="complement"
-                autoComplete="complement"
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
+            <Grid item xs={12}>
               <TextField
                 required
                 fullWidth
                 name="password"
-                label="Senha"
+                label="Password"
                 type="password"
                 id="password"
                 autoComplete="new-password"
               />
             </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                required
-                fullWidth
-                name="passwordAgain"
-                label="Repetir Senha"
-                type="passwordAgain"
-                id="passwordAgain"
-                autoComplete="new-passwordAgain"
+            <Grid item xs={12}>
+              <FormControlLabel
+                control={<Checkbox value="allowExtraEmails" color="primary" />}
+                label="I want to receive inspiration, marketing promotions and updates via email."
               />
             </Grid>
           </Grid>
           <Button type="submit" fullWidth variant="contained" sx={{ mt: 3, mb: 2 }}>
-            Cadastre-se
+            Sign Up
           </Button>
-          <Grid container justifyContent="center">
+          <Grid container justifyContent="flex-end">
             <Grid item>
-              <Link href="/Site/signIn" variant="body2" style={{ textDecoration: 'none' }}>
-                Já tenho uma conta
+              <Link href="/Site/signIn" variant="body2">
+                Already have an account? Sign in
               </Link>
             </Grid>
           </Grid>
