+++ src/pages/_app.tsx
@@ -1,33 +1,23 @@
 import * as React from 'react';
-import type { ReactElement, ReactNode } from 'react'
 import Head from 'next/head';
 import { AppProps } from 'next/app';
-import type { NextPage } from 'next'
 import { ThemeProvider } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { CacheProvider, EmotionCache } from '@emotion/react';
-import theme from 'constants/theme';
-import createEmotionCache from 'constants/createEmotionCache';
-import DefaultLayout from 'Components/Layout/default';
+import theme from '../Constants/theme';
+import createEmotionCache from '../Constants/createEmotionCache';
+import { Header } from '../Components/Header';
+import { Footer } from '../Components/Footer';
 
 // Client-side cache, shared for the whole session of the user in the browser.
 const clientSideEmotionCache = createEmotionCache();
 
-type NextPageWithLayout = NextPage & {
-  getLayout?: (page: ReactElement) => ReactNode
-}
-
-type MyAppProps = AppProps & {
-  Component: NextPageWithLayout
+interface MyAppProps extends AppProps {
   emotionCache?: EmotionCache;
 }
 
 export default function MyApp(props: MyAppProps) {
   const { Component, emotionCache = clientSideEmotionCache, pageProps } = props;
-
-  // Use the layout defined at the page level, if available
-  const getLayout = Component.getLayout ?? ((page) => <DefaultLayout>{page}</DefaultLayout>)
-
   return (
     <CacheProvider value={emotionCache}>
       <Head>
@@ -37,7 +27,9 @@ export default function MyApp(props: MyAppProps) {
       <ThemeProvider theme={theme}>
         {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
         <CssBaseline />
-        {getLayout(<Component {...pageProps} />)}
+        <Header/>
+        <Component {...pageProps} />
+        <Footer/>
       </ThemeProvider>
     </CacheProvider>
   );
