--- src/styles/app.module.scss
@@ -1,9 +1,0 @@
-.container {
-  display: flex;
-  height: 100vh;
-
-  main{
-    height: 100vh;
-    width: 100%;
-  }
-}
