--- src/styles/global.scss
@@ -1,23 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  background-color: #f2f2f2;
-  height: auto;
-  min-height: 100vh;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  font-family: Poppins, sans-serif;
-}
-
-p,
-span,
-button,
-div {
-  font-family: Roboto, sans-serif;
-}
