--- src/styles/pages/index.module.scss
@@ -1,150 +1,0 @@
-.bannerDetails {
-  background-color: rgba(247, 248, 250, 0.8);
-  width: 730px;
-  height: 350px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  box-sizing: border-box;
-  padding: 4rem;
-
-  img {
-    width: 300px;
-  }
-
-  p {
-    font: 500 1rem Roboto, sans-serif;
-    text-align: center;
-    padding: 0 6rem;
-  }
-
-  button {
-    background-color: #006D77;
-    width: 20rem;
-    padding: 0.8rem 2rem;
-    border-radius: 25rem;
-    font: 800 1rem Roboto, sans-serif;
-    color: #FFFFFF;
-    border: 0px solid #000000;
-  }
-}
-
-.cooworkDetails {
-  padding: 2rem 2rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-
-  article {
-    display: flex;
-    justify-content: space-around;
-    margin-bottom: 2rem;
-  }
-}
-
-.cooworkDetailsHeader {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  padding: 2rem 0;
-
-  h1 {
-    margin: 0;
-    font: 500 2rem Poppins, sans-serif;
-  }
-
-  p {
-    margin: 0;
-    margin-bottom: 2rem;
-    color: #5E6E89;
-    letter-spacing: 1px;
-  }
-}
-
-.cooworkDetailsCard {
-  background-color: #EDF6F9;
-  padding: 2rem;
-  width: 17rem;
-  height: 17rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  box-shadow: 2px 2px 14px #a8a8a8;
-
-  img {
-    width: 120px;
-  }
-
-  h3 {
-    font-size: 1rem;
-    text-align: center;
-  }
-
-  span {
-    text-align: center;
-  }
-}
-
-.cooworkRanking {
-  padding: 2rem 2rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  background-color: #E5F2EB;
-
-  article {
-    display: flex;
-    justify-content: space-around;
-    margin-bottom: 2rem;
-  }
-}
-
-.cooworkRankingHeader {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  padding: 2rem 0;
-
-  h1 {
-    margin: 0;
-    font: 500 2rem Poppins, sans-serif;
-  }
-
-  p {
-    margin: 0;
-    margin-bottom: 2rem;
-    color: #5E6E89;
-    letter-spacing: 1px;
-  }
-}
-
-.cooworkRankingCard {
-  background-color: #F7F8FA;
-  padding: 2rem;
-  width: 17rem;
-  height: 17rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  box-shadow: 2px 2px 14px #a8a8a8;
-
-  img {
-    width: 120px;
-  }
-
-  h3 {
-    font-size: 1rem;
-    text-align: center;
-    font-weight: 500;
-  }
-
-  span {
-    text-align: center;
-    color: #7C8087;
-  }
-}
