+++ tsconfig.json
@@ -1,13 +1,5 @@
 {
   "compilerOptions": {
-    "baseUrl": "./src",
-    "paths": {
-      "Common/*": ["components/common/*"],
-      "Site-components/*": ["components/Site/*"],
-      "Dash-components/*": ["components/Dashboard/*"],
-      "Components/*": ["components/*"],
-      "@Book": ["pages/Dashboard/book"]
-    },
     "allowJs": true,
     "allowSyntheticDefaultImports": true,
     "jsx": "preserve",
@@ -39,5 +31,5 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
+  ]
 }
